GoPro.DEVC

Description

DEVC is one of the GPMF data types that contains the bulk of interesting telemetry data from within GPMF streams. This module doesn't currently provide high level access to *all* DEVC data (some of it remains low level), but it currently has useful representations of things that seemed interesting to the author.

Documentation

mkDEVC :: FourCC -> [Value] -> DEVC

Given a FourCC value (specifically, DEVC) and a list of Values, produce a DEVC value.

data DEVC

Constructors

DEVC

Fields

_dev_id :: Int

_dev_name :: String

_dev_telems :: Map String Telemetry

dev_id :: Lens' DEVC Int

dev_name :: Lens' DEVC String

dev_telems :: Lens' DEVC (Map String Telemetry)

data Accelerometer

Constructors

Accelerometer

Fields

_acc_temp :: Float

_acc_vals :: [(Float, Float, Float)]

acc_temp :: Lens' Accelerometer Float

acc_vals :: Lens' Accelerometer [(Float, Float, Float)]

data Gyroscope

Constructors

Gyroscope

Fields

_gyro_temp :: Float

_gyro_vals :: [(Float, Float, Float)]

gyro_temp :: Lens' Gyroscope Float

gyro_vals :: Lens' Gyroscope [(Float, Float, Float)]

data Face

Constructors

Face

Fields

_face_id :: Int

_face_x :: Float

_face_y :: Float

_face_w :: Float

_face_h :: Float

_face_smile :: Float

face_id :: Lens' Face Int

face_x :: Lens' Face Float

face_y :: Lens' Face Float

face_w :: Lens' Face Float

face_h :: Lens' Face Float

face_smile :: Lens' Face Float

data GPSReading

Constructors

GPSReading

Fields

_gpsr_lat :: Double

_gpsr_lon :: Double

_gpsr_alt :: Double

_gpsr_speed2d :: Double

_gpsr_speed3d :: Double

gpsr_lat :: Lens' GPSReading Double

gpsr_lon :: Lens' GPSReading Double

gpsr_alt :: Lens' GPSReading Double

gpsr_speed2d :: Lens' GPSReading Double

gpsr_speed3d :: Lens' GPSReading Double

data GPS

Constructors

GPS

Fields

_gps_p :: Int

_gps_time :: UTCTime

_gps_readings :: [GPSReading]

gps_p :: Lens' GPS Int

gps_time :: Lens' GPS UTCTime

gps_readings :: Lens' GPS [GPSReading]

data AudioLevel

Constructors

AudioLevel

Fields

_audio_rms :: [Int]

_audio_peak :: [Int]

audio_rms :: Lens' AudioLevel [Int]

audio_peak :: Lens' AudioLevel [Int]

data Location

Constructors

Snow

Urban

Indoor

Water

Vegetation

Beach

_Snow :: Prism' Location ()

_Urban :: Prism' Location ()

_Indoor :: Prism' Location ()

_Water :: Prism' Location ()

_Vegetation :: Prism' Location ()

_Beach :: Prism' Location ()

data TVals

Constructors

TVUnknown [Value]

TVAccl Accelerometer

TVGyro Gyroscope

TVFaces [Face]

TVGPS GPS

TVAudioLevel AudioLevel

TVScene [Map Location Float]

_TVUnknown :: Prism' TVals [Value]

_TVAccl :: Prism' TVals Accelerometer

_TVGyro :: Prism' TVals Gyroscope

_TVFaces :: Prism' TVals [Face]

_TVGPS :: Prism' TVals GPS

_TVAudioLevel :: Prism' TVals AudioLevel

_TVScene :: Prism' TVals [Map Location Float]

data Telemetry

Constructors

Telemetry

Fields

_tele_stmp :: Word64

_tele_tsmp :: Int

_tele_name :: String

_tele_values :: TVals

tele_stmp :: Lens' Telemetry Word64

tele_tsmp :: Lens' Telemetry Int

tele_name :: Lens' Telemetry String

tele_values :: Lens' Telemetry TVals